Platform-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
docker==6.1.3
pydantic==2.5.0

